RSS Hyperlinks Dilemma
Trying to figure out how to get the hyperlinks in my RSS list view feed
to display using the ChildBrowser plugin.  main.js file is responsible
for the feed and I have been able to get the link that take the user to
the blog but can't get links that the blog feeds the device to utilize
the CB plugin.